+++ Source/SwiftLintCore/Extensions/Configuration+Parsing.swift
@@ -27,6 +27,7 @@ extension Configuration {
     // MARK: - Initializers
     /// Creates a Configuration value based on the specified parameters.
     ///
+    /// - parameter parentConfiguration:    The parent configuration, if any.
     /// - parameter dict:                   The untyped dictionary to serve as the input for this typed configuration.
     ///                                     Typically generated from a YAML-formatted file.
     /// - parameter ruleList:               The list of rules to be available to this configuration.
@@ -34,6 +35,7 @@ extension Configuration {
     ///                                     settings in `dict`.
     /// - parameter cachePath:              The location of the persisted cache on disk.
     public init(
+        parentConfiguration: Configuration? = nil,
         dict: [String: Any],
         ruleList: RuleList = RuleRegistry.shared.list,
         enableAllRules: Bool = false,
@@ -75,7 +77,10 @@ extension Configuration {
         )
 
         Self.validateConfiguredRulesAreEnabled(
-            configurationDictionary: dict, ruleList: ruleList, rulesMode: rulesMode
+            parentConfiguration: parentConfiguration,
+            configurationDictionary: dict,
+            ruleList: ruleList,
+            rulesMode: rulesMode
         )
 
         self.init(
@@ -147,35 +152,115 @@ extension Configuration {
     }
 
     private static func validateConfiguredRulesAreEnabled(
+        parentConfiguration: Configuration?,
         configurationDictionary dict: [String: Any],
         ruleList: RuleList,
         rulesMode: RulesMode
     ) {
         for key in dict.keys where !validGlobalKeys.contains(key) {
             guard let identifier = ruleList.identifier(for: key),
-                let rule = ruleList.list[identifier] else {
+                let ruleType = ruleList.list[identifier] else {
                     continue
             }
-
-            let message = "Found a configuration for '\(identifier)' rule"
 
             switch rulesMode {
             case .allEnabled:
                 return
-
             case .only(let onlyRules):
-                if Set(onlyRules).isDisjoint(with: rule.description.allIdentifiers) {
-                    Issue.genericWarning("\(message), but it is not present on '\(Key.onlyRules.rawValue)'.").print()
-                }
-
+                let issue = validateConfiguredRuleIsEnabled(onlyRules: onlyRules, ruleType: ruleType)
+                issue?.print()
             case let .default(disabled: disabledRules, optIn: optInRules):
-                if rule is any OptInRule.Type, Set(optInRules).isDisjoint(with: rule.description.allIdentifiers) {
-                    Issue.genericWarning("\(message), but it is not enabled on '\(Key.optInRules.rawValue)'.").print()
-                } else if Set(disabledRules).isSuperset(of: rule.description.allIdentifiers) {
-                    Issue.genericWarning("\(message), but it is disabled on '\(Key.disabledRules.rawValue)'.").print()
-                }
+                let issue = validateConfiguredRuleIsEnabled(
+                    parentConfiguration: parentConfiguration,
+                    disabledRules: disabledRules,
+                    optInRules: optInRules,
+                    ruleType: ruleType
+                )
+                issue?.print()
             }
         }
+    }
+
+    static func validateConfiguredRuleIsEnabled(
+        parentConfiguration: Configuration?,
+        disabledRules: Set<String>,
+        optInRules: Set<String>,
+        ruleType: any Rule.Type
+    ) -> Issue? {
+        var enabledInParentRules: Set<String> = []
+        var disabledInParentRules: Set<String> = []
+        var allEnabledRules: Set<String> = []
+
+        if case .only(let onlyRules) = parentConfiguration?.rulesMode {
+            enabledInParentRules = onlyRules
+        } else if case .default(let parentDisabledRules, let parentOptInRules) = parentConfiguration?.rulesMode {
+            enabledInParentRules = parentOptInRules
+            disabledInParentRules = parentDisabledRules
+        }
+        allEnabledRules = enabledInParentRules
+            .subtracting(disabledInParentRules)
+            .union(optInRules)
+            .subtracting(disabledRules)
+
+        return validateConfiguredRuleIsEnabled(
+            parentConfiguration: parentConfiguration,
+            enabledInParentRules: enabledInParentRules,
+            disabledInParentRules: disabledInParentRules,
+            disabledRules: disabledRules,
+            optInRules: optInRules,
+            allEnabledRules: allEnabledRules,
+            ruleType: ruleType
+        )
+    }
+
+    static func validateConfiguredRuleIsEnabled(
+        onlyRules: Set<String>,
+        ruleType: any Rule.Type
+    ) -> Issue? {
+        if onlyRules.isDisjoint(with: ruleType.description.allIdentifiers) {
+            return Issue.ruleNotPresentInOnlyRules(ruleID: ruleType.identifier)
+        }
+        return nil
+    }
+
+    // swiftlint:disable:next function_parameter_count
+    static func validateConfiguredRuleIsEnabled(
+        parentConfiguration: Configuration?,
+        enabledInParentRules: Set<String>,
+        disabledInParentRules: Set<String>,
+        disabledRules: Set<String>,
+        optInRules: Set<String>,
+        allEnabledRules: Set<String>,
+        ruleType: any Rule.Type
+    ) -> Issue? {
+        if case .allEnabled = parentConfiguration?.rulesMode {
+            if disabledRules.contains(ruleType.identifier) {
+                return Issue.ruleDisabledInDisabledRules(ruleID: ruleType.identifier)
+            }
+            return nil
+        }
+
+        let allIdentifiers = ruleType.description.allIdentifiers
+
+        if allEnabledRules.isDisjoint(with: allIdentifiers) {
+            if !disabledRules.isDisjoint(with: allIdentifiers) {
+                return Issue.ruleDisabledInDisabledRules(ruleID: ruleType.identifier)
+            }
+            if !disabledInParentRules.isDisjoint(with: allIdentifiers) {
+                return Issue.ruleDisabledInParentConfiguration(ruleID: ruleType.identifier)
+            }
+
+            if ruleType is any OptInRule.Type {
+                if enabledInParentRules.union(optInRules).isDisjoint(with: allIdentifiers) {
+                    return Issue.ruleNotEnabledInOptInRules(ruleID: ruleType.identifier)
+                }
+            } else if case .only(let enabledInParentRules) = parentConfiguration?.rulesMode,
+                      enabledInParentRules.isDisjoint(with: allIdentifiers) {
+                return Issue.ruleNotEnabledInParentOnlyRules(ruleID: ruleType.identifier)
+            }
+        }
+
+        return nil
     }
 
     private static func warnAboutMisplacedAnalyzerRules(optInRules: [String], ruleList: RuleList) {